% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateWRSobj.R
\name{CreateWRSobj}
\alias{CreateWRSobj}
\title{Create WRS Object}
\usage{
CreateWRSobj(nA, nB, alternative = c("two.sided", "less", "greater"),
  inclPermMat = T)
}
\arguments{
\item{nA}{the number of sample in the first group}

\item{nB}{the number of sample in the secend group}

\item{inclPermMat}{a logical vector indicating TRUE when include a permutation matrix, FALSE otherwise.}
}
\description{
This function generates WRS object to feed in later test function
}
\examples{
CreateWRSobj(7,7)

}

