% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TopK.R
\name{plot.TopK}
\alias{plot.TopK}
\title{Plot triangle figure for a TopK object}
\usage{
\method{plot}{TopK}(x, WobsDX = 1, ...)
}
\arguments{
\item{x}{\code{TopK} object, typically result of TopK.test}

\item{WobsDX}{the index of observed data in all permutations. Default value is \code{1}}

\item{...}{other parameters to be passed through to plotting functions.}
}
\description{
A triangle plot illustrate the results of TopK test. This is a plot only for the scenario of \code{K=2}.
The x-axis denotes the \code{-log10 p_(1)} -value and the y-axis denotes \code{-log10 p_(2)} -value.
The p-values were tabulated across all permutations and displayed as counts (for WRS test);
as scattered and smoothed points (for permutation t-test).
The red circle denotes the value combination that was observed in the actual data.
}

