% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.TopK.R
\name{print.TopK}
\alias{print.TopK}
\title{Print Methods for TopK test}
\usage{
\method{print}{TopK}(x, ...)
}
\arguments{
\item{x}{\code{TopK} object, typically result of TopK.test.}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
Pringting objects of class \code{"TopK"}.
}

