% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.TopK.R
\name{summary.TopK}
\alias{summary.TopK}
\title{Summarizing the TopK test results}
\usage{
\method{summary}{TopK}(object, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}

\item{x}{\code{TopK} object, typically result of TopK.test}
}
\description{
\code{summary} method for class \code{"lm"}.
}

