% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopK_PermT.test.R
\name{TopK_PermT.test}
\alias{TopK_PermT.test}
\title{TopK exact test based on permutation T p-values}
\usage{
TopK_PermT.test(X, nA, nB, Kvals = c(1, 2, 3, 4, 5, 10, 25), B = 0,
  ReturnType = "list", vrb = T)
}
\arguments{
\item{X}{a data matrix containing features as rows and samples as columns.}

\item{nA}{number of samples in group 1}

\item{nB}{number of samples in group 2}

\item{Kvals}{a numeric vector indicating how many "K" we choose.}

\item{B}{the number of inner permutation, default is "0" for exact test.}

\item{ReturnType}{character(1) specify how to return the results.
Could be "list" or "vector". See detail.}

\item{vrb}{a logical vector indicating TRUE when show some words while running, FALSE otherwise.}
}
\description{
This function perform TopK exact test on a two group data set (Case-Control study).
The univariate p-values based on permutation t-test.
}
\examples{
TopK_PermT.test(X,7,7)

}

