% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopK_PermT.test.R
\name{TopK_PermT.test}
\alias{TopK_PermT.test}
\title{TopK exact test based on permutation T p-values}
\usage{
TopK_PermT.test(X, nA, nB, Kvals = c(1, 2, 3, 4, 5, 10, 25), B = 0,
  alternative = c("two.sided", "less", "greater"), ties.method = c("random",
  "min", "max", "average"), pval = FALSE, ReturnType = "TopK", vrb = T)
}
\arguments{
\item{X}{a data matrix containing features as rows and samples as columns.}

\item{nA}{number of samples in group 1}

\item{nB}{number of samples in group 2}

\item{Kvals}{a numeric vector indicating how many \code{"K"} we choose.}

\item{B}{the number of inner permutation, default is "0" for exact test.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}.
You can specify just the initial letter.}

\item{ties.method}{a character string specifying how ties are treated, see ‘Details’; can be abbreviated.}

\item{pval}{If TRUE, the permutation matrix will be populated by p-values from t-test.
If FALSE, the permutation matrix will be directly populated by t statistics (negative absolute value).}

\item{ReturnType}{character(1) specify how to return the results.
Could be \code{"list"}, \code{"vector"} or rich format \code{"TopK"}. See detail.}

\item{vrb}{a logical vector indicating TRUE when show some words while running, FALSE otherwise.}
}
\description{
This function perform TopK exact test on a two group data set (Case-Control study).
The univariate p-values based on permutation t-test.
}
\examples{
TopK_PermT.test(X,7,7)

}

